Rename project & fix minor comments
