+++ inngest-spring-boot-demo/src/test/java/com/inngest/springbootdemo/InngestControllerTest.java
@@ -1,10 +1,9 @@
-package com.inngest.springdemo;
+package com.inngest.springbootdemo;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.ResultActions;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
