+++ inngest-spring-boot-demo/src/test/java/com/inngest/springbootdemo/SpringBootDemoApplicationTests.java
@@ -1,10 +1,10 @@
-package com.inngest.springdemo;
+package com.inngest.springbootdemo;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class SpringDemoApplicationTests {
+class SpringBootDemoApplicationTests {
 
 	@Test
 	void contextLoads() {
